files-22 (#25)
* [file-service] use deleteDirectory in case it's a directory

* [file-service] throw CannotDeleteFile if doesn\'t exists

* [test] delete directory with files inside of it

+++ src/Http/Services/FileService.php
@@ -90,7 +90,15 @@ class FileService implements FileServiceContract
             $path = $url;
         }
         try {
-            $deleted = $this->disk->delete($path);
+            if ($this->disk->exists($path)) {
+                if ($this->disk->mimeType($path) === 'directory') {
+                    $deleted = $this->disk->deleteDirectory($path);
+                } else {
+                    $deleted = $this->disk->delete($path);
+                }
+            } else {
+                $deleted = false;
+            }
             if (!$deleted) {
                 throw new CannotDeleteFile($url);
             }
